pyproject.toml
[project]
name = "ember"
version = "0.1.0.dev0"
description = "A foundation model for agents that live."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Slow Lit Labs" }]
keywords = ["llm", "agents", "fine-tuning", "evals"]

dependencies = [
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "rich>=13.7.0",
    "tenacity>=9.0.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
training = [
    "torch>=2.4.0",
    "transformers>=4.45.0",
    "peft>=0.13.0",
    "datasets>=3.0.0",
    "accelerate>=1.0.0",
    "axolotl>=0.5.0",
]
inference = [
    "vllm>=0.6.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[project.urls]
Homepage = "https://slowlitlabs.co"
Repository = "https://github.com/slowlitlabs/ember"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests", "evals"]
python_files = ["test_*.py", "*_test.py"]
